Server/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
executing==2.2.1
fastapi==0.121.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
littleutils==0.2.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-smi==0.1.3
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
polars==1.35.1
polars-runtime-32==1.35.1
psutil==7.1.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sorcery==0.2.2
SQLAlchemy==2.0.44
starlette==0.49.3
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.3.225
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.0
